Feature/new firmware move failure classes (#29)
* Move udp client failures to calling class
* Move CRC mismatch of ScannerReply to ScannerReply
* Move MonitoringFrame format error to Monitoring frame
* Move inline code to class definition
* Move StringStreamFailure to raw_processing namespace
* Incorporate review feedback
