+++ test/unit_tests/unittest_monitoring_frame_msg.cpp
@@ -21,10 +21,8 @@
 
 #include "psen_scan_v2/angle_conversions.h"
 #include "psen_scan_v2/istring_stream_builder.h"
-#include "psen_scan_v2/monitoring_frame_format_error.h"
 #include "psen_scan_v2/monitoring_frame_msg.h"
 #include "psen_scan_v2/raw_processing.h"
-#include "psen_scan_v2/string_stream_failure.h"
 #include "psen_scan_v2/udp_frame_dumps.h"
 #include "psen_scan_v2/raw_data_array_conversion.h"
 
@@ -66,7 +64,7 @@ TEST(FieldHeaderTest, testReadHeaderTooShortFailure)
   builder.add(too_short_header);
   std::istringstream is{ builder.get() };
 
-  EXPECT_THROW(MonitoringFrameMsg::readFieldHeader(is);, StringStreamFailure);
+  EXPECT_THROW(MonitoringFrameMsg::readFieldHeader(is);, raw_processing::StringStreamFailure);
 }
 
 class MonitoringFrameMsgTest : public ::testing::Test
@@ -120,7 +118,8 @@ TEST_F(MonitoringFrameMsgTest, testReadScanCounterInvalidLengthFailure)
   std::istringstream is{ builder.get() };
 
   uint32_t scan_counter;
-  EXPECT_THROW(MonitoringFrameMsg::readScanCounter(is, scan_counter, length);, MonitoringFrameFormatError);
+  EXPECT_THROW(MonitoringFrameMsg::readScanCounter(is, scan_counter, length);
+               , MonitoringFrameMsg::MonitoringFrameFormatError);
 }
 
 TEST_F(MonitoringFrameMsgTest, testReadScanCounterMissingPayloadFailure)
@@ -128,7 +127,7 @@ TEST_F(MonitoringFrameMsgTest, testReadScanCounterMissingPayloadFailure)
   const uint16_t length = 4;
   std::istringstream is;
   uint32_t scan_counter;
-  EXPECT_THROW(MonitoringFrameMsg::readScanCounter(is, scan_counter, length);, StringStreamFailure);
+  EXPECT_THROW(MonitoringFrameMsg::readScanCounter(is, scan_counter, length);, raw_processing::StringStreamFailure);
 }
 
 TEST_F(MonitoringFrameMsgTest, testReadMeasuresSuccess)
@@ -146,7 +145,7 @@ TEST_F(MonitoringFrameMsgTest, testReadMeasuresMissingPayloadFailure)
   const uint16_t length = 2 * expected_measures_.size();
   std::istringstream is;
   std::vector<double> measures;
-  EXPECT_THROW(MonitoringFrameMsg::readMeasures(is, measures, length);, StringStreamFailure);
+  EXPECT_THROW(MonitoringFrameMsg::readMeasures(is, measures, length);, raw_processing::StringStreamFailure);
 }
 
 TEST_F(MonitoringFrameMsgTest, testReadMeasuresTooMuchMeasures)
@@ -164,7 +163,7 @@ TEST_F(MonitoringFrameMsgTest, testReadMeasuresTooFewMeasures)
   const uint16_t length = 2 * (expected_measures_.size() + 1);
   std::istringstream is = buildExpectedMeasuresStream();
   std::vector<double> measures;
-  EXPECT_THROW(MonitoringFrameMsg::readMeasures(is, measures, length);, StringStreamFailure);
+  EXPECT_THROW(MonitoringFrameMsg::readMeasures(is, measures, length);, raw_processing::StringStreamFailure);
 }
 
 class MonitoringFrameMsgFromRawTest : public ::testing::Test
@@ -227,7 +226,7 @@ TEST_F(MonitoringFrameMsgFromRawTest, testUnknownFieldId)
   const auto raw_frame_data = convertToMaxSizeRawData(test_data.hex_dump);
 
   MonitoringFrameMsg msg;
-  EXPECT_THROW(msg = MonitoringFrameMsg::fromRawData(raw_frame_data);, MonitoringFrameFormatError);
+  EXPECT_THROW(msg = MonitoringFrameMsg::fromRawData(raw_frame_data);, MonitoringFrameMsg::MonitoringFrameFormatError);
 }
 
 TEST_F(MonitoringFrameMsgFromRawTest, testTooLargeFieldLength)
@@ -236,7 +235,7 @@ TEST_F(MonitoringFrameMsgFromRawTest, testTooLargeFieldLength)
   const auto raw_frame_data = convertToMaxSizeRawData(test_data.hex_dump);
 
   MonitoringFrameMsg msg;
-  EXPECT_THROW(msg = MonitoringFrameMsg::fromRawData(raw_frame_data);, MonitoringFrameFormatError);
+  EXPECT_THROW(msg = MonitoringFrameMsg::fromRawData(raw_frame_data);, MonitoringFrameMsg::MonitoringFrameFormatError);
 }
 
 }  // namespace psen_scan_v2
